mueve logica impresion y arregla detalles
